# Swagger-Artline-1.0.0-resolved.json
{
  "swagger": "2.0",
  "info": {
    "description": "Este proyecto consiste en el desarrollo de una API para una red social enfocada al ámbito artístico. Se busca que esta API tenga acceso a una base de datos de artistas y obras, para así crear una serie de endpoints enfocados a la creación, consulta y edición de dicha información. Todo esto con el fin de crear una comunidad de usuarios que disfrutan de ver y compartir el arte.\n",
    "version": "1.0.0",
    "title": "Artline",
    "contact": {
      "email": "it@artline.com"
    }
  },
  "host": "virtserver.swaggerhub.com",
  "basePath": "/miuclara91/Artline/1.0.0",
  "tags": [
    {
      "name": "usuarios",
      "description": "Operaciones en los usuarios"
    },
    {
      "name": "comentarios",
      "description": "Operaciones en los comentarios"
    },
    {
      "name": "publicaciones",
      "description": "Operaciones de las publicaciones"
    }
  ],
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/usuarios": {
      "post": {
        "tags": [
          "usuarios"
        ],
        "summary": "Creación de un nuevo usuario",
        "operationId": "createUsuario",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "usuario",
            "description": "Objeto usuario que se creará para insertarse en la red social artline",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Usuario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario creado"
          },
          "500": {
            "description": "Error fatal"
          }
        }
      }
    },
    "/usuarios/{id}": {
      "put": {
        "tags": [
          "usuarios"
        ],
        "summary": "Modificación de un usuario por su id",
        "operationId": "updateUsuario",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ObjectId del usuario que será modificado",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "usuario",
            "description": "Objeto usuario que se modificará en la red social artline",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Usuario"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usuario modificado"
          },
          "500": {
            "description": "Error fatal"
          }
        }
      }
    }
  },
  "definitions": {
    "Usuario": {
      "type": "object",
      "required": [
        "bio",
        "username",
        "password"
      ],
      "properties": {
        "password": {
          "type": "string",
          "format": "password"
        },
        "username": {
          "type": "string",
          "example": "Juanita"
        },
        "followercount": {
          "type": "integer",
          "format": "int64"
        },
        "bio": {
          "type": "string",
          "example": "Mi nombres es Juanita"
        },
        "postcount": {
          "type": "integer",
          "format": "int64"
        },
        "likes": {
          "type": "integer",
          "format": "int64"
        }
      }
    }
  }
}